### 2024-06-10 #6862 @editorialbot
[PRE REVIEW]: SetVis: Visualizing Large Numbers of Sets and Intersections
**Submitting author:** @layik (<a href="http://orcid.org/0000-0003-1912-4890">Layik Hama</a>)
**Repository:** https://github.com/alan-turing-institute/setvis
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.1
