### 2024-06-10 #6862 @sappelhoff
@crvernon thanks for the invite -- unfortunately I have to decline at this time, as I have too many open commitments.
