### 2024-06-11 #6862 @layik
Hi @mstimberg. Thank you for agreeing to manage the review.

> If you have any suggestions for potential reviewers, then please mention them here

No particular person jumped out I am afraid, from this filter:

https://reviewers.joss.theoj.org/reviewers/search?language=26&area_id=5&keywords=Visualization+&name=&commit=Search

Good luck with the chase. Thank you.
