### 2024-06-17 #6862 @mstimberg
@jnothman I did not quickly find your email address, so I'll ping you here directly :blush: Would you be willing to review this submission for JOSS – your expertise from working on UpSetPlot would be very much appreciated.  We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. To find out more about JOSS in general, please see https://joss.theoj.org/about.
