### 2024-06-24 #6862 @mstimberg
Thanks for helping us out!

PS: Since this is a new GitHub account which does not have much information: this is the account of Jean-Baptiste Lamy, researcher at the [LIMICS](http://www.limics.fr/?page_id=697&lang=en) lab at University Paris 13 (France), working on information and knowledge visualization (among other things). Very happy to have you on board!
