### 2024-06-25 #6862 @jibalamy
> 👉📄 [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.06862/joss.06862/10.21105.joss.06862.pdf) 📄 [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.06862/joss.06862/10.21105.joss.06862.pdf) 📄 👈

These links are dead, is it normal?
