### 2024-06-25 #6862 @mstimberg
> These links are dead, is it normal?

The documents generated as part of the pre-review are no longer available once the review issue (#6925) has been opened. You can find the latest version over there : https://github.com/openjournals/joss-reviews/issues/6925#issuecomment-2186432043
