### 2024-06-28 #6925 @layik
Hi

RE one of the general checks where the checkbox states:

> If the paper contains original results, results are entirely reproducible by reviewers

Just to be proactive here, I am not sure if I include the relevant files (including notebooks) the reviewers would be able to reproduce the results of the tech-evaluation due to the heavy computation required. I have done the tech-evaluation to be reproducible. I would happily push them to the branch or upload them here if needed. Likewise would happily provide the output of the evaluation (csv files) that was used to generate the plots in the paper if that is deemed sufficient for the check.
