### 2024-07-13 #6925 @mstimberg
@layik Apologies for the late reply on this topic. Yes, please make the files to run the tech-evaluation available in some form. While reviewers might not be able/willing to run the full computation to verify the exact results, they might want to have a look at the benchmarking procedure itself. Ideally, the code would be easily adaptable for them to run a reduced version (e.g. just the first points) of the benchmark.
