### 2024-07-14 #6925 @jibalamy
I add some difficulties for installing the module, but it's now Ok.
It is not available on Python 3.12, which is the default version on my Linux distribution (Arch Linux). Then running the tutorial requires Bokeh, but in a old version (it works with 2.4 but not with the latest version). Why the most recent version of Python and Bokeh are not supported? This would simplify the installation of the module.
