### 2024-07-14 #6925 @jibalamy
Then I tried the tutorial 1, but I am stuck at step 5, "Making a selection". I am supposed to select rows with missing combinations, but these rows are not contiguous. I tried to select them with control or shift + mouse selection, but I am unable to select non contiguous rows.
Additionally, the combination heatmap is difficult to read, because there are column labels but no row labels. Is that normal?
