### 2024-07-15 aef7d6d
Add technical evaluation papers for joss review as requested
