+++ joss/tech-evaluation/utils.py
@@ -1,0 +1,390 @@
+import math
+
+import pandas as pd
+import numpy as np
+import itertools
+import matplotlib.pyplot as plt
+import upsetplot
+import collections.abc
+from setvis.membership import Membership
+from setvis.plots import PlotSession
+
+
+def generate_general_missing(num_rows, num_cols, num_combinations):
+    """Generate a data frame that has columns of type 'object', and a general pattern of missing data.
+    Each row will be missing at least one value.
+    Each column to be missing by itself at least once.
+    There to be at least one intersection of each possible length (2 - num_cols).
+    Each combination will occur an equal number of times, to the extent that is possible
+    The rows are shuffled so they are in a 'random' order.
+
+
+    Parameters
+    ----------
+    num_cols : int
+        The number of columns in the data frame
+    num_rows : int
+        The number of rows in the data frame
+    num_combinations : int
+        The number of combinations of missing values (num_rows >= num_combinations; num_combinations >= 2 * num_cols - 1)
+
+    Returns
+    -------
+    pd.DataFrame
+        Data frame containing the dataset.
+
+    """
+    # Calculate the maximum number of combinations of each length (1 - num_cols)
+    max_comb = []
+    # Create a list of lists that store [combination length, max possible combinations]
+    length_ncomb = []
+    for l1 in range(num_cols):
+        max_comb.append(math.comb(num_cols, l1 + 1))
+
+        if l1 > 0 and l1 < num_cols - 1:
+            # Don't calculate this for combination length = 1 or num_cols
+            length_ncomb.append([l1 + 1, max_comb[-1]])
+
+    # Sort the list of lists
+    length_ncomb.sort(key=lambda x: x[1])
+
+    if (
+        num_combinations <= sum(max_comb)
+        and num_rows >= num_combinations
+        and num_combinations >= num_cols + (num_cols - 1)
+    ):
+        # Calculate the number of combinations of each length (1 column, 2 columns, etc.) so the distribution is as even as possible
+        # If you're interested 'esi' is an acronym for exclusive set intersections
+        esi_params = {1: num_cols, num_cols: 1}
+        ncombs = esi_params[1] + esi_params[num_cols]
+        # Loop over the other combination lengths, leaving the one that can have the most combinations until last
+        cnt = 0
+        for item in length_ncomb:
+            clen = item[0]
+            esi_params[clen] = min(
+                item[1],
+                int(
+                    (num_combinations - ncombs + len(length_ncomb) - cnt - 1)
+                    / (len(length_ncomb) - cnt)
+                ),
+            )
+            ncombs = ncombs + esi_params[clen]
+            cnt = cnt + 1
+
+        intersections_template = np.ones(
+            shape=(num_combinations, num_cols)
+        )  # matrix of ones
+
+        # Loop over the parameter dictionary to create a list of lists. Each
+        # item is a combination of missing values
+        ii = 0
+        for key, value in esi_params.items():
+            # Randomise the order of the sets for ESIs with this cardinality
+            columns = np.arange(1, num_cols + 1)
+            # The number of ESIs and transactions with this cardinality
+            esi_count = 0
+            # Loop until the specified number of ESIs has been generated
+            for i in itertools.combinations(columns, key):
+                for n in i:
+                    intersections_template[ii][n - 1] = np.nan
+
+                # print(intersections_template[ii, :])
+                ii = ii + 1
+                esi_count = esi_count + 1
+                if esi_count == value:
+                    break
+
+        # Create a data frame that contains all the intersections, repeating them until there are the specified number of rows
+        rows = []
+
+        count = 0
[…]
