### 2024-07-15 #6925 @layik
Dear @mstimberg and @jibalamy. Please see the files added to the `joss` branch of the repository. I intend to remove this from the repository once the review is concluded if you do not object to it.
https://github.com/alan-turing-institute/setvis/commit/aef7d6dcf03703827a460bfe4b8fa9d1f364de98

As for
> Why the most recent version of Python and Bokeh are not supported?

The answer is because we decided to restrict it to the versions being tested within the CI setup. Happy to relax this condition but must discuss with other authors.

> I am stuck at step 5, "Making a selection". I am supposed to select rows with missing combinations, but these rows are not contiguous. I tried to select them with control or shift + mouse selection, but I am unable to select non contiguous rows.

I am going to ask for some help here from @ots22 if not will get you clarification from other authors offline and post it here.

> Additionally, the combination heatmap is difficult to read, because there are column labels but no row labels. Is that normal?

Again, I will ask @ots22 but again will get some answers offline if not.

Thank you very much for your time.
