+++ pyproject.toml
@@ -25,7 +25,7 @@ classifiers = [
 ]
 
 [tool.poetry.dependencies]
-python = ">=3.8,<=3.12"
+python = ">=3.8,<3.13"
 bokeh = "^2.3"
 numpy = "^1.21"
 pandas = "^1.3"
