### 2024-07-16 #6925 @ots22
Hi @jibalamy,

Thank you for the review so far.

> > Why the most recent version of Python and Bokeh are not supported?

> The answer is because we decided to restrict it to the versions being tested within the CI setup. Happy to relax this condition but must discuss with other authors.

I think we can remove the upper bound on the Python version now; I'll submit a PR. We can control the versions we test in CI independently and we can add 3.12 there too.

I haven't tried bumping the Bokeh version, but something we can check.

> >  I am stuck at step 5, "Making a selection". I am supposed to select rows with missing combinations, but these rows are not contiguous. I tried to select them with control or shift + mouse selection, but I am unable to select non contiguous rows.

> I am going to ask for some help here from @ots22 if not will get you clarification from other authors offline and post it here.

It should work by holding shift with a mouse selection, which is just the default behaviour of the [box select tool](https://docs.bokeh.org/en/latest/docs/user_guide/interaction/tools.html#boxselecttool) in Bokeh.

Do you have the same problem with multiple selections on [this example](https://docs.bokeh.org/en/latest/docs/examples/interaction/linking/data_table_plot.html) from the Bokeh gallery?

> > Additionally, the combination heatmap is difficult to read, because there are column labels but no row labels. Is that normal?

> Again, I will ask @ots22 but again will get some answers offline if not.

Each row visually represents a particular combination of present and missing values.  There are as many rows as there are distinct combinations and these are shown one above the other as the rows.  There isn't a meaningful row label that could be included here - the particular pattern of boxes is everything there is to know.
