### 2024-07-17 37ee99c
Use python 3.12 in automated testing workflow
