+++ .github/workflows/main.yml
@@ -30,13 +30,11 @@ jobs:
       if: steps.cache-deps.outputs.cache-hit != 'true'
       run: |
         python -m pip install --upgrade pip
-        pip install poetry
-        poetry update
-        poetry install -E all
+        python -m pip install -e '.[all]'
 
     - name: Tests
       run: |
-        poetry run pytest
+        python -m pytest
 
     - name: Notebooks [Linux]
       if: runner.os == 'Linux' && matrix.python-version == '3.12'
