### 2024-07-17 2e486e7
Remove redundant pip install of the package (workflow)
