+++ .github/workflows/main.yml
@@ -39,5 +39,4 @@ jobs:
     - name: Notebooks [Linux]
       if: runner.os == 'Linux' && matrix.python-version == '3.12'
       run: |
-        pip install -e '.[all]'
         cd tests && bash check_notebooks.sh
