### 2024-07-19 78fc3e1
Bokeh 3 compatibility (still works with Bokeh 2)
