### 2024-07-19 a0c557d
Switch to psycopg2-binary
