+++ pyproject.toml
@@ -53,7 +53,7 @@ doc = ["Sphinx", "pydata-sphinx-theme"]
 test = ["pytest"]
 
 performance-extras = ["numexpr", "Bottleneck"]
-db = ["psycopg2"]
+db = ["psycopg2-binary"]
 
 extra = [
   "notebook",
@@ -73,7 +73,7 @@ all = [
   "pytest",
   "numexpr",
   "Bottleneck",
-  "psycopg2",
+  "psycopg2-binary",
 ]
 
 [build-system]
