+++ pyproject.toml
@@ -26,20 +26,21 @@ classifiers = [
 
 [tool.poetry.dependencies]
 python = ">=3.8"
-bokeh = "^2.3"
-numpy = "^1.21"
-pandas = "^1.3"
-pydantic = "^1.8"
+bokeh = ">=2.3,<4"
+numpy = ">=1.21,<3"
+pandas = ">=1.3,<3"
+pydantic = ">=1.8,<3"
 tomli = "^2.0"
+setuptools = ">=68.1"
 
 notebook = { version = "^6.4", optional = true }
-matplotlib = { version = "^3.4.3", optional = true }
+matplotlib = { version = "^3.4", optional = true }
 scipy = { version = "^1.7", optional = true }
-scikit-learn = { version = ">=0.2", optional = true }
-Sphinx = { version = "^4.3", optional = true }
+scikit-learn = { version = ">=0.2,<2", optional = true }
+Sphinx = { version = ">=4.3", optional = true }
 pydata-sphinx-theme = { version = "^0.7", optional = true }
-pytest = { version = "^6.2", optional = true }
-psycopg2 = { version = "^2.9", optional = true }
+pytest = { version = ">=6.2", optional = true }
+psycopg2-binary = { version = "^2.9", optional = true }
 numexpr = { version = "^2.7", optional = true }
 Bottleneck = { version = "^1.3", optional = true }
 
