+++ pyproject.toml
@@ -26,7 +26,7 @@ classifiers = [
 
 [tool.poetry.dependencies]
 python = ">=3.8"
-bokeh = ">=2.3,<4"
+bokeh = "^3.0"
 numpy = ">=1.21,<3"
 pandas = ">=1.3,<3"
 pydantic = ">=1.8,<3"
