### 2024-07-24 f14d195
Upgrade to Bokeh 3: Panel replaced with TabPanel
