### 2024-07-24 e7ea730
Bokeh figure size option
Use stretch_width due to change in behaviour of stretch_both
