### 2024-07-24 cbfeef6
Use restricted box selection (Bokeh plot tools)
