### 2024-07-24 3414c82
Relax required version of notebook
