### 2024-07-26 a5b79d4
Increment minor version of Bokeh dependency
