### 2024-07-26 #6925 @layik
Hi @kirangadhave, @mstimberg and others. Thank you so much for your reviews so far. Oliver (@ots22) has done a deps review and taken into account all your input. The PR is ready to merge and he will release a minor release ASAP.

https://github.com/alan-turing-institute/setvis/pull/133

Can I please ask you to re-run your selection tests making sure you have bokeh 3.4? You can use the branch in the above PR or you can wait for the release.

I believe this would address most of the issues raised.

As for the tech evaluation, please let me know if you need me to do anything else.

One more time thank you very much for taking time to review the submission.
