+++ pyproject.toml
@@ -25,7 +25,7 @@ classifiers = [
 ]
 
 [tool.poetry.dependencies]
-python = ">=3.8"
+python = ">=3.9"
 bokeh = "^3.4"
 numpy = ">=1.21,<3"
 pandas = ">=1.3,<3"
