+++ .github/workflows/main.yml
@@ -14,7 +14,7 @@ jobs:
 
     strategy:
       matrix:
-        python-version: ['3.8', '3.11', '3.12']
+        python-version: ['3.9', '3.11', '3.12']
         os: [windows-latest, ubuntu-latest]
 
     steps:
