### 2024-07-26 e600f2c
Remove reference to old test configs in docs
