### 2024-07-26 f95c409
Update name of conda environment in installation instructions
