### 2024-07-26 1e663ab
Update conda instructions
