### 2024-07-26 164add3
Fix name of optional dependency set (notebooks) in installation instructions
