### 2024-07-26 #6925 @ots22
Hi @kirangadhave and @mstimberg,

As @layik noted, we have now upgraded to Bokeh 3 and relaxed the required versions of some other packages which we think might help with the selection tool misbehaving.

Please could you upgrade setvis to the latest version, either from GitHub or PyPI (the latter should be version 0.1rc4) and see if this has fixed the problem? You should not need to manually upgrade Bokeh or the other dependencies, although you might want to check that you have ended up with bokeh >= 3.4 if you are reusing an environment.
