### 2024-08-14 #6925 @jibalamy
Hi, I successfully tested the Github version of SetVis with Python 3.12.4 and Bokeh 3.5.1.
I think I now understand how to read the heatmap; indeed it works somehow similarly to Upset, but with color instead of bars. I think you should describe better that adaptation/change; after reading the manuscript, I was thinking that Setvis was producing visualization similar to Upset, and then I was puzzled when I see the demo because the visualization was actually not the same. Possibly, it can be a good idea to display a simple dataset both with Upset and Setvis, for facilitating the understanding and the comparison.
