### 2024-08-14 #6925 @jibalamy
I now have difficulties to understand the various bar charts available in the other tabs. I looked at "Example - import data to visualize sets". In the "Value bar chart", it appears that set cardinalities range from 4 to 32. But in the "Value count histogram", set cardinalities range from 1 to 11... I cannot understand what is represented on these bars. Similarly, I cannot understand the other bar charts (some have decimal cardinality, such as 0.5 or 1.5, which is surprising... probably, 0.5-1.5 actually means 1 ?). Is there some explanations of these bar charts, ideally on the simple example ?
