### 2024-08-14 #6925 @jibalamy
I also obtained several warnings (see below) when running the tests with pytest, is it normal?

tests/test_plots.py::test_IntersectionBarChart
  /home/jiba/src/setvis/setvis/membership.py:43: DeprecationWarning: `in1d` is deprecated. Use `np.isin` instead.
    return list(np.array(universe)[~np.in1d(universe, selection)])

tests/test_plots.py::test_PlotSession
  /usr/lib/python3.12/site-packages/tornado/ioloop.py:274: DeprecationWarning: There is no current event loop
    loop = asyncio.get_event_loop()

tests/test_plots.py::test_PlotSession
  /usr/lib/python3.12/site-packages/pydantic/main.py:1087: PydanticDeprecatedSince20: The `dict` method is deprecated; use `model_dump` instead. Deprecated in Pydantic V2.0 to be removed in V3.0. See Pydantic V2 Migration Guide at https://errors.pydantic.dev/2.8/migration/
    warnings.warn('The `dict` method is deprecated; use `model_dump` instead.', category=PydanticDeprecatedSince20)

tests/test_plots.py::test_SetBarChart
  /home/jiba/src/setvis/setvis/plots.py:220: DeprecationWarning: `in1d` is deprecated. Use `np.isin` instead.
    np.in1d(col_labels, selection.columns, invert=True)

tests/test_plots.py::test_SetCardinalityHistogram
  /home/jiba/src/setvis/setvis/plots.py:351: DeprecationWarning: `in1d` is deprecated. Use `np.isin` instead.
    np.in1d(col_labels, selection.columns, invert=True)

tests/test_plots.py::test_IntersectionBarChart
  /home/jiba/src/setvis/setvis/plots.py:501: DeprecationWarning: `in1d` is deprecated. Use `np.isin` instead.
    np.in1d(
