### 2024-08-22 #6925 @kirangadhave
@ots22 @layik I am having trouble installing and running the new package version (`0.1rc4`) in Python 3.12. I am creating a new virtual environment for this test.

I tried installing it from PyPi and Github.

### PyPi
```pip install 'setvis[notebook]'```

### Github
```pip install git+pip install git+https://github.com/alan-turing-institute/setvis@dependency-versions```

Both commands also install `bokeh==3.5.1`

I get the following error message when I try to import the `PlotSession` class following the `Tutorial 1` instructions:

![setvis_bokeh_3 5](https://github.com/user-attachments/assets/175670f9-1cf1-4556-9d7b-898c18391f4a)

I tried downgrading Bokeh to 3.4 manually and still got the same message.

![setvis_bokeh_3 4](https://github.com/user-attachments/assets/f0977269-4f7f-465e-9be6-28b16fb72328)

Can you add a _"review"_ Python environment to the `joss` branch in the SetVis repo? You can share the `requirements.txt` file generated by the `pip freeze` command inside a working environment. Feel free to let me know if you need anything from me to fix the issue.
