### 2024-08-23 #6925 @ots22
Hi @kirangadhave,

The extra dependency set required for the notebooks is in fact called 'notebook**s**' (with an s), not 'notebook'. Please could you try the installation with `pip install setvis[notebooks]`?  Pip gives a warning but it is easy to miss.

The instructions did have the 'notebook' typo at some point although I think it is now corrected. If you came across this elsewhere please could you let me know where and I can correct it there too?

I'm not sure if that is the issue here, but it might result in the wrong kernel getting used by jupyter for instance.

The import error you get is surprising because setvis 0.1rc4 does not import Panel any more (this was a change needed for Bokeh 3, it was renamed TabPanel, [link to source](https://github.com/alan-turing-institute/setvis/blob/v0.1rc4/setvis/plots.py#L12)), so I wonder if an old version of setvis is getting used somehow? Please could you check the output of `%pip show setvis` (that is, `%pip` rather than `!pip`)?

> Can you add a "review" Python environment to the joss branch in the SetVis repo? You can share the requirements.txt file generated by the pip freeze command inside a working environment.

That sounds like a good idea - done. It's named `requirements_notebooks.txt` and has a pinned set of versions based on the same 'notebooks' extras (if you want to use the postgres functionality we can generate a new one with 'db').
