### 2024-08-23 #6925 @ots22
If none of the above helps, please could you send me your own `pip freeze` output and I'll try to investigate?
