### 2024-08-23 #6925 @ots22
Hi @jibalamy,

> I also obtained several warnings (see below) when running the tests with pytest, is it normal?
> ...

These are all innocuous so please ignore them for now. I will push a fix at some point.

I hope we can reply to your other comments soon.
