### 2024-09-03 c32dbbf
paper.md: minor based on reviewer comments
