### 2024-09-03 #6925 @layik
Minor change from me for now, based on the agreed plan mentioned above.
https://github.com/alan-turing-institute/setvis/commit/c32dbbf84bf898bc8354a67bde6ad4f832a620d8
