### 2024-09-13 #6925 @layik
Hi @mstimberg and team. Again, I should have communicated the time estimate. We agreed on "early Oct". Hope that is not too far away now.
