### 2024-09-13 #6925 @layik
It would be great to know if the reviews (so far) finished ticking the checklist boxes? For example I can see that @jibalamy has ticked the automated test box but @kirangadhave has not. I would happily include the "unticked" ones in the current plan once I know the reviewers have finished their "ticking".
