### 2024-09-26 #6925 @mstimberg
> Hi @mstimberg and team. Again, I should have communicated the time estimate. We agreed on "early Oct". Hope that is not too far away now. Apologies for the long delays and thank you very much for your patience.

Apologies for not replying earlier. Yes, early October is acceptable, but please try to not delay things further, since it always takes additional effort for reviewers to come back to a paper/software after a long break.
