### 2024-09-26 #6925 @mstimberg
> It would be great to know if the reviews (so far) finished ticking the checklist boxes? For example I can see that @jibalamy has ticked the automated test box but @kirangadhave has not. I would happily include the "unticked" ones in the current plan once I know the reviewers have finished their "ticking". Thanks again!

@jibalamy, @kirangadhave could you please reply to this question, i.e. are your unticked boxes up-to-date with the comments you made here earlier, or do they correspond to criticism/suggestions you did not put in writing yet (or are some of them obsolete, i.e. can be checked now). Thanks alot :pray:
