### 2024-09-27 #6925 @layik
Hello again
I just sat down to track our response and gladly saw your replies.

> since it always takes additional effort for reviewers to come back to a paper/software after a long break
Absolutely and I can only apologise to everyone. I will push this forward including the latest comments by @jibalamy (thank you so much for taking the time especially outside Jupyter).

I know we still need a confirmation of ticking box state by @kirangadhave.

I will keep this thread under constant watch.
