### 2024-09-27 alan-turing-institute/setvis#135 @layik
JOSS review issues
Reviewer (Kiran) has finalised his review [status](https://github.com/openjournals/joss-reviews/issues/6925#issuecomment-2231778546).

Matters outstanding:

**Docs**

Unticked by [Kiran](https://github.com/openjournals/joss-reviews/issues/6925#issuecomment-2231778546)

- [x] Community guidelines (addressed via various commits)

- [x] agreed in "early Oct" plan
- [x] [summary](https://github.com/openjournals/joss-reviews/issues/6925#issuecomment-2188151095) which is ticked off by
[Kiran](https://github.com/openjournals/joss-reviews/issues/6925#issuecomment-2231778546)

Then latest **comments** by jibalamy
- [x] Community guidelines are very limited in the documentation, and probably need to be enriched (e.g. how to submit bugs or propose help? etc)
- [x] Regarding the paper, it needs more details (and screenshots) to present the various graphics (as I said in a previous message)
- [x] I also tried to run SetVis without Jupyter, but I did not succeed. There is no documentation about that, and actually it is not clear whether it is possible or not. This point should also be clarified. (_I think you can do this better than me @ots22, thank you)_

Latest [comments](https://github.com/openjournals/joss-reviews/issues/6925#issuecomment-2414170816) by Kiran:
- [x] "A few issue templates on github for bugs, features, etc"
- [x] "A way to contact the devs for support (email or discussions on Github)." package is under Alan Turing Institute account + developer github identities visible.
- [x] "A section to the README for setting up local dev environment, testing, etc." see bold section on README
