### 2024-09-27 #6925 @layik
link https://github.com/alan-turing-institute/setvis/issues/135
