### 2024-09-29 #6925 @layik
Dear @jibalamy, I know the delay in getting this is not helping, as shown above, I am going through the issue on the repo and while addressing this comment:

> Performance claim is difficult to assess without a proper benchmark, but I did not find any benchmark in the code

I have committed all the code for the bench-marking claim in the paper in the `joss` branch. Had you seen [these](https://github.com/alan-turing-institute/setvis/tree/joss/joss/tech-evaluation) files? Thank you.
