+++ docs/source/index.rst
@@ -31,9 +31,10 @@ Contents
    :maxdepth: 2
 
    installation
+   api_reference
    tutorials
    documentation
-   api_reference
+   contributing
 
 Introductory Example
 ====================
