### 2024-09-29 alan-turing-institute/setvis#135 @layik
Response draft:
- summarise actions in "early Oct" plan
- new contribute.rst in doc, tired up with the guide for generating the docs
- tbc
