+++ notebooks/README.md
@@ -21,13 +21,13 @@ The setvis package includes the following notebooks:
 There are six visualizations in setvis. These are described using [Example - import data to visualize missingness](https://github.com/alan-turing-institute/setvis/blob/main/notebooks/Example%20-%20import%20data%20to%20visualize%20missingness.ipynb).
 ![value bar chart](images/value_bar_chart.JPG)
 <p align="center">The value bar chart shows the number of missing values (Y axis) in each field (X axis).</p>
-
+![combination heatmap](images/combination_heatmap.JPG)
 <p align="center">The combination heatmap shows the dataset’s fields (X axis), each combination of missing values (Y axis) and the number of records that are in each combination (colour).</p>
-
+![combination bar chart](images/combination_bar_chart.JPG)
 <p align="center">The combination bar chart shows the number of records (Y axis) in each combination of missing values (X axis).</p>
-
+![value count histogram](value_count_histogram.JPG)
 <p align="center">The value count histogram shows the distribution of the number of missing values (X axis) and number of fields (Y-axis). This helps setvis scale to datasets with many fields.</p>
-
+![combination count histogram](combination_count_histogram.JPG)
 <p align="center">The combination count histogram shows the distribution of the number of records (X axis) and number of combinations of missing values (Y-axis). This helps setvis scale to datasets with many fields.</p>
-
+![combination length histogram](combination_length_histogram.JPG)
 <p align="center">The combination length histogram shows the distribution of the length (X axis) and number of combinations of missing values (Y-axis). This helps setvis scale to datasets with many fields.</p>
