+++ notebooks/README.md
@@ -28,11 +28,11 @@ There are six visualizations in setvis. These are described using [Example - imp
 ![combination bar chart](images/combination_bar_chart.JPG)
 <p align="center">The combination bar chart shows the number of records (Y axis) in each combination of missing values (X axis).</p>
 
-![value count histogram](value_count_histogram.JPG)
+![value count histogram](images/value_count_histogram.JPG)
 <p align="center">The value count histogram shows the distribution of the number of missing values (X axis) and number of fields (Y-axis). This helps setvis scale to datasets with many fields.</p>
 
-![combination count histogram](combination_count_histogram.JPG)
+![combination count histogram](images/combination_count_histogram.JPG)
 <p align="center">The combination count histogram shows the distribution of the number of records (X axis) and number of combinations of missing values (Y-axis). This helps setvis scale to datasets with many fields.</p>
 
-![combination length histogram](combination_length_histogram.JPG)
+![combination length histogram](images/combination_length_histogram.JPG)
 <p align="center">The combination length histogram shows the distribution of the length (X axis) and number of combinations of missing values (Y-axis). This helps setvis scale to datasets with many fields.</p>
