+++ notebooks/README.md
@@ -20,19 +20,19 @@ The setvis package includes the following notebooks:
 # Visualizations
 There are six visualizations in setvis. These are described using [Example - import data to visualize missingness](https://github.com/alan-turing-institute/setvis/blob/main/notebooks/Example%20-%20import%20data%20to%20visualize%20missingness.ipynb).
 ![value bar chart](images/value_bar_chart.JPG)
-<p align="center">The value bar chart shows the number of missing values (Y axis) in each field (X axis).</p>
+<p align="center">The value bar chart shows the number of missing values (Y axis) in each field (X axis). The staircase of bars suggests that the DIAG_02 - DIAG_19 fields have a monotone pattern of missingness.</p>
 
 ![combination heatmap](images/combination_heatmap.JPG)
-<p align="center">The combination heatmap shows the dataset’s fields (X axis), each combination of missing values (Y axis) and the number of records that are in each combination (colour).</p>
+<p align="center">The combination heatmap shows the dataset’s fields (X axis), each combination of missing values (Y axis) and the number of records that are in each combination (colour). The top, 4th from top and bottom six combinations are a monotone pattern. However, the other seven combinations show that there is another pattern that has gaps in the DIAG fields.</p>
 
 ![combination bar chart](images/combination_bar_chart.JPG)
-<p align="center">The combination bar chart shows the number of records (Y axis) in each combination of missing values (X axis).</p>
+<p align="center">The combination bar chart shows the number of records (Y axis) in each combination of missing values (X axis). There are 1 - 40 records in each combination.</p>
 
 ![value count histogram](images/value_count_histogram.JPG)
-<p align="center">The value count histogram shows the distribution of the number of missing values (X axis) and number of fields (Y-axis). This helps setvis scale to datasets with many fields.</p>
+<p align="center">The value count histogram shows the distribution of the number of missing values (X axis) and number of fields (Y-axis). This histogram setvis scale to datasets with many fields.</p>
 
 ![combination count histogram](images/combination_count_histogram.JPG)
-<p align="center">The combination count histogram shows the distribution of the number of records (X axis) and number of combinations of missing values (Y-axis). This helps setvis scale to datasets with many fields.</p>
+<p align="center">The combination count histogram shows the distribution of the number of records (X axis) and number of combinations of missing values (Y-axis). This histogram setvis scale to datasets with many fields.</p>
 
 ![combination length histogram](images/combination_length_histogram.JPG)
-<p align="center">The combination length histogram shows the distribution of the length (X axis) and number of combinations of missing values (Y-axis). This helps setvis scale to datasets with many fields.</p>
+<p align="center">The combination length histogram shows the distribution of the length (X axis) and number of combinations of missing values (Y-axis). This histogram setvis scale to datasets with many fields.</p>
