### 2024-10-11 alan-turing-institute/setvis#135 @layik
Runnig in iPython using `reprexpy`, I believe has been addressed via #133

```python
from setvis.plots import PlotSession
#> Traceback (most recent call last):
#> Cell In[3], line 1
#> ----> 1 from setvis.plots import PlotSession
#> File ~/code/python/setvis/setvis/plots.py:12
#>      10 from bokeh.palettes import Oranges256
#>      11 from bokeh.transform import linear_cmap
#> ---> 12 from bokeh.models.widgets import Panel, Tabs
#>      13 from bokeh.events import SelectionGeometry
#>      14 import bokeh.io
#> ImportError: cannot import name 'Panel' from 'bokeh.models.widgets' (/home/layik/miniconda3/envs/3point9/lib/python3.9/site-packages/bokeh/models/widgets/__init__.py)

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

from sklearn import tree, preprocessing
#> Traceback (most recent call last):
#> Cell In[7], line 1
#> ----> 1 from sklearn import tree, preprocessing
#> ModuleNotFoundError: No module named 'sklearn'
from sklearn.metrics import mutual_info_score
#> Traceback (most recent call last):
#> Cell In[8], line 1
#> ----> 1 from sklearn.metrics import mutual_info_score
#> ModuleNotFoundError: No module named 'sklearn'
df = pd.read_csv("../examples/datasets/Synthetic_APC_DIAG_Fields.csv")
df.tail()
#>      DIAG_01  DIAG_02  DIAG_03  DIAG_04  DIAG_05  DIAG_06  DIAG_07  DIAG_08  \
195        1      NaN      NaN      NaN      NaN      NaN      NaN      NaN
196        1      NaN      NaN      NaN      NaN      NaN      NaN      NaN
197        1      NaN      NaN      NaN      NaN      NaN      NaN      NaN
198        1      NaN      NaN      NaN      NaN      NaN      NaN      NaN
199        1      NaN      NaN      NaN      NaN      NaN      NaN      NaN

     DIAG_09  DIAG_10  ADMIAGE  ADMIMETH  Mortality PROCODE3  SEX
195      NaN      NaN        8        65          0      abc    1
196      NaN      NaN       37       111          0      kkk    1
197      NaN      NaN       10        40          0      fff    2
198      NaN      NaN       12        65          0      abc    1
199      NaN      NaN       10       111          0      ggg    2
session = PlotSession(df)
#> Traceback (most recent call last):
#> Cell In[11], line 1
#> ----> 1 session = PlotSession(df)
#> NameError: name 'PlotSession' is not defined
session.add_plot(name="all data")
#> Traceback (most recent call last):
#> Cell In[12], line 1
#> ----> 1 session.add_plot(name="all data")
#> NameError: name 'session' is not defined
```
