### 2024-10-15 alan-turing-institute/setvis#135 @ots22
> I also tried to run SetVis without Jupyter, but I did not succeed. There is no documentation about that, and actually it is not clear whether it is possible or not. This point should also be clarified. (I think you can do this better than me @ots22, thank you)

I think it is reasonable to say that jupyter is the preferred way of using setvis.  We depend on Bokeh for interactive figures so it's hard to replicate the workflow without something like jupyter.

Still, it is possible to start a Bokeh server and return the port, which opens up some possibilities for non-interactive functionality.  That idea is contained in [this example](https://github.com/alan-turing-institute/setvis/blob/main/notebooks/Example%20-%20plotting%20outside%20the%20notebook.ipynb) (fourth cell down), but there isn't much detail on it.

I have been working through the steps of using this and can put a note in the docs: both how to do it if desired, but also that it is 'advanced' usage.
