### 2024-10-15 #6925 @kirangadhave
I was able to complete my review over the weekend. I was able to install setvis and test it out in a new environment correctly. I was able to run benchmarks as well.

From the checklist the only thing I could not find was the `Community guidelines`. I would recommend adding the following:
- A section to the `README` for setting up local dev environment, testing, etc.
- A few issue templates on github for `bugs`, `features`, etc.
- A way to contact the devs for support (email or discussions on Github).

With the final addition, I can mark off the last item on the review checklist.
