### 2024-10-15 1117c95
Remove box_zoom and pan tools from SetBarChart for consistency with other plots
