### 2024-10-15 fd727a7
Add log message after Bokeh server start; import inside conditional
