### 2024-10-16 alan-turing-institute/setvis#135 @layik
Thank you very much @ots22, I have just updated the issue inline with the updates in the review thread [latest](https://github.com/openjournals/joss-reviews/issues/6925#issuecomment-2414170816), sending an email now.
