### 2024-10-16 84d2fea
create issue templates
As per JOSS review recommendations
