### 2024-10-16 alan-turing-institute/setvis#135 @ots22
PR https://github.com/alan-turing-institute/setvis/pull/140 addresses the comment that it is unclear whether use outside Jupyter is supported
