### 2024-10-17 alan-turing-institute/setvis#135 @layik
Thank you @ots22, once #141 is merged, can we please draft a release before I send my review response full draft to the reviewers? Thank you.
