### 2024-10-17 alan-turing-institute/setvis#135 @ots22
New release candidate: [GitHub](https://github.com/alan-turing-institute/setvis/releases/tag/v0.1rc5), [PyPI](https://pypi.org/project/setvis/)
