### 2024-10-18 fc7321d
paper.mnd: address review comment in summary #135
