+++ joss/paper.md
@@ -31,7 +31,7 @@ affiliations:
 
 # Summary
 
-Set-type data occurs in many domains (e.g., life sciences [@lamy2019rainbio], health [@landolfi2022screening] and the retail industry [@adnan2018set]), as well as generic applications such as analysing structures of missing data [@ruddle2022using] and association rule mining [@wang2020visual]. SetVis is new matrix-based set visualization software, implemented as a Python package which is available from PyPi https://pypi.org/project/setvis. SetVis uses a memory-efficient design, operates with datasets held in RAM or a database such as PostgreSQL, and allows user interaction graphically or programmatically. A technical evaluation shows that SetVis uses orders of magnitude less memory than UpSet plot [@lex2014upset] python package [@nothman2022] implementation.
+Set-type data occurs in many domains (e.g., life sciences [@lamy2019rainbio], health [@landolfi2022screening] and the retail industry [@adnan2018set]), as well as generic applications such as analysing structures of missing data [@ruddle2022using] and association rule mining [@wang2020visual]. SetVis is new matrix-based set visualization software, implemented as a Python package which is available from PyPi https://pypi.org/project/setvis. The documentation is available from https://setvis.readthedocs.io which contains various hands-on Jupyter notebooks guiding users to use the package. SetVis uses a memory-efficient design, operates with datasets held in RAM or a database such as PostgreSQL, and allows user interaction graphically or programmatically. A technical evaluation shows that SetVis uses orders of magnitude less memory than UpSet plot [@lex2014upset] python package [@nothman2022] implementation.
 
 # Statement of need
 Although a wide variety of set visualization software has been developed [@jia2021venn;@alsallakh2016powerset], most of such software generates Venn or Euler diagrams so is only suitable for data that contain fewer than 10 sets [@jia2021venn]. Other software visualizes 50+ sets but either has little support for set intersection tasks [@alper2011design;@dork2012pivotpaths;@kim2007visualizing;@freiler2008interactive] or only visualizes pairwise intersections [@molbiotools2022;@yalcin2015aggreset].
