### 2024-10-18 #6925 @layik
Dear JOSS reviewers, my final response is circulated for internal check but will be pasting it here either end of today or Monday morning (21st October). We are grateful for your contributions: @mstimberg, @kirangadhave and @jibalamy.
