### 2024-10-21 #6925 @layik
Dear @mstimberg (and @jibalamy and @kirangadhave),

Since the submission, we have addressed various aspects of the package, the paper and the claims within it. The technical evaluation claims within the paper have now been reproduced and confirmed by both reviewers. We have made some minor changes within the paper inline with the contributions from the two reviewers. One of these is the [comment](https://github.com/openjournals/joss-reviews/issues/6925#issuecomment-2376805000) asking for more details of the package and graphics to present it. We are within the word limit but have added a generic sentence pointing readers to the RTD (Read The Docs) pages as we know the package will grow, therefore keeping the paper ever green is important.

In term of the source code, the main issue addressed has been the dependency on Bokeh which is now resolved in the main branch. The main area picked up on by the reviewers has been documentation and "community" guidelines, we have made various changes including guiding users and collaborators to use the package and contribute to it such as bug report template added to the package. A substantial update has gone into the tutorial notebooks which is pointed to from the RTD pages of setvis.

The joss branch which includes the submission ("paper.md" etc) files remains to be merged (for your review) and we are open and ready to publish the benchmarking and other files related to the review ([PR#138](https://eur03.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Falan-turing-institute%2Fsetvis%2Fpull%2F138&data=05%7C02%7CL.Hama%40leeds.ac.uk%7C5f1069eff0fe4f09094708dcef4f9468%7Cbdeaeda8c81d45ce863e5232a535b7cb%7C0%7C0%7C638648372278630707%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C0%7C%7C%7C&sdata=f4hxXoBGEp1OpLrwv9Pxp38F35lzC0eE%2Fz2xaMNTmXM%3D&reserved=0)) but have sought your input as already linked in this [issue](https://eur03.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Fopenjournals%2Fjoss-reviews%2Fissues%2F6925&data=05%7C02%7CL.Hama%40leeds.ac.uk%7C5f1069eff0fe4f09094708dcef4f9468%7Cbdeaeda8c81d45ce863e5232a535b7cb%7C0%7C0%7C638648372278686253%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C0%7C%7C%7C&sdata=BYpLJ00ykIgdzZ04t44RACAhKabTre2F2rF1lWPdQlk%3D&reserved=0).

A new package version with above changes (except joss branch) was released ["v0.1rc5"](https://github.com/alan-turing-institute/setvis/releases/tag/v0.1rc5) on 17thOctober 2024. For details of the changes please see issue [#135](https://github.com/alan-turing-institute/setvis/issues/135), [PR#138](https://github.com/alan-turing-institute/setvis/pull/138), release v0.1rc5 and the history (commits) on both branches main and joss since the submission on 24th June 2024.

We (Prof Ruddle, Dr Strickson and I) are grateful for your contributions and welcome any further comments or reviews before the paper is accepted.
