### 2024-10-22 #6925 @jibalamy
I've read the revised paper, but one of my remarks, about the absence of screenshot of the proposed tool has not been addressed. Is it normal ?
