### 2024-10-22 #6925 @layik
Just to confirm, for that comment we decided to add this sentence "The documentation is available from https://setvis.readthedocs.io which contains various hands-on Jupyter notebooks guiding users to use the package." where there are screenshots. I am happy to address **quickly** if this is a requirement.
