### 2024-10-22 #6925 @jibalamy
Yes, if there is enough space in the paper, I think that a screenshot of the proposed tool would make it much more explicit.
