### 2024-10-23 8cc833f
paper: add example screenshot in design section #135
