+++ joss/paper.md
@@ -39,7 +39,9 @@ Although a wide variety of set visualization software has been developed [@jia20
 The best-known software for analysing rich patterns in set data are the R and Python UpSet plot packages [@conway2017upsetr;@nothman2022], but the memory requirement of both packages increases linearly with the number of cells (i.e., rows $\times$ columns) in a dataset, which makes the packages unusable with big data. The ACE software [@ruddle2022ace] uses more memory-efficient data structures, but first requires the whole of a dataset to have been loaded into RAM (again, clearly an issue for big data), and is a stand-alone Java application that cannot be integrated with Jupyter Notebooks or similar workflows. The SetVis python package addresses the above collective weaknesses because it: (a) operates with datasets that may be either held in RAM or out of core (in a PostgreSQL database), (b) stores sets and intersections in memory-efficient data structures (like ACE), (c) can be used within Jupyter Notebooks (or similar) to aid the replicability of analysis workflows, and (d) allows users to interact graphically in a notebook as well as programmatically.
 
 # Design
-SetVis provides the same six built-in visualizations as ACE [@ruddle2022ace]. The main two show visualizations of sets (in a bar chart) and set intersections (in a heatmap). The other four visualizations make SetVis scalable to data that contains large numbers of sets and/or intersections, by showing histograms of set cardinality, intersection degree and intersection cardinality, and an intersection degree bar chart. All of the visualizations are interactive (implemented with Bokeh [@bokeh2018]), but users may also interact programmatically and freely interleave the two forms of interaction. Examples and tutorials are provided with the installation.
+![An example APC combination heatmap shows the fields (X axis), each combination of missing values (Y axis) and the number of records that are in each combination (colour) of the APC (Admitted Patient Care) dataset included in the package. The top, 4th from top and bottom six combinations are a monotone pattern. However, the other seven combinations show that there is another pattern that has gaps in the DIAG fields. \label{fig:heatmap}](../notebooks/images/combination_heatmap.JPG)
+
+SetVis provides the same six built-in visualizations as ACE [@ruddle2022ace]. The main two show visualizations of sets (in a bar chart) and set intersections (in a heatmap). The other four visualizations make SetVis scalable to data that contains large numbers of sets and/or intersections, by showing histograms of set cardinality, intersection degree and intersection cardinality, and an intersection degree bar chart. All of the visualizations are interactive (implemented with Bokeh [@bokeh2018]), but users may also interact programmatically and freely interleave the two forms of interaction. Examples and tutorials are provided with the installation. A screenshot of SetVis version `v0.1rc5` of the heatmap visualizations within Jupyter notebooks is shown in Figure \ref{fig:heatmap}.
 
 Jupyter notebooks have been widely adopted in the Python data science ecosystem for exploratory data analysis. It is considered good practice for computational notebooks to obey principles of (i) top-to-bottom re-executability and (ii) repeatability, including by others [@quaranta2022notebooks]. The SetVis design allows these principles to be respected.
 
