### 2024-10-23 #6925 @layik
We are within the 1000 limit if not over, the bot I think counts the "code" words. I thought I should make the changes and give you the option @jibalamy. Please see the change in this commit or regenerate the PDF.

https://github.com/alan-turing-institute/setvis/commit/8cc833f134ec24dcd2c5e71ef72f007d62d5a926

Let me know if that (going over the 1000 word limit) is acceptable otherwise happy to reset head of the branch.
