### 2024-10-23 #6925 @mstimberg
Hi @layik and @jibalamy – just to chime in on this question: it is ultimately up to you whether you decide to include the screenshot in the paper or not, you do not need to worry about the word count in that regard. The 1000-word limit is not strict (but of course please do not go much over it), and as you noticed, the measurement is rather inexact.
