### 2024-10-24 #6925 @layik
We would prefer to go with ever-green paper and let the plots grow/change  with the package. That is no screenshot in the paper but a reference to the docs.
>  One of these is the https://github.com/openjournals/joss-reviews/issues/6925#issuecomment-2376805000 asking for more details of the package and graphics to present it. We are within the word limit but have added a generic sentence pointing readers to the RTD (Read The Docs) pages as we know the package will grow, therefore keeping the paper ever green is important.
