### 2024-10-24 #6925 @jibalamy
Hello, I largely prefer the version with the screenshot, it helps to understand the visualization and it makes the paper "self-contained" without relying on external links (which may be broken, or not available, e.g. if printing the paper).
