### 2024-10-24 #6925 @kirangadhave
I agree with @jibalamy on keeping the screenshot.

I also saw the contribution instructions you added. I’ll check the item in review.
