### 2024-10-29 #6925 @layik
Thank you @mstimberg. Here is the DOI of a Zenodo archive of the release
https://doi.org/10.5281/zenodo.14006398

> Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

The title is different as the package is called "SetVis" but the paper is called "SetVis: Visualizing Large Numbers of Sets and Intersections" everything else is checked and done.
