### 2024-10-29 e0d8085
Remove AAS fields for paper header
