### 2024-10-29 7f87ab4
Spell out country in affiliations
