+++ joss/paper.md
@@ -29,17 +29,17 @@ affiliations:
 
 # Summary
 
-Set-type data occurs in many domains (e.g., life sciences [@lamy2019rainbio], health [@landolfi2022screening] and the retail industry [@adnan2018set]), as well as generic applications such as analysing structures of missing data [@ruddle2022using] and association rule mining [@wang2020visual]. SetVis is new matrix-based set visualization software, implemented as a Python package which is available from PyPi https://pypi.org/project/setvis. The documentation is available from https://setvis.readthedocs.io which contains various hands-on Jupyter notebooks guiding users to use the package. SetVis uses a memory-efficient design, operates with datasets held in RAM or a database such as PostgreSQL, and allows user interaction graphically or programmatically. A technical evaluation shows that SetVis uses orders of magnitude less memory than UpSet plot [@lex2014upset] python package [@nothman2022] implementation.
+Set-type data occurs in many domains such as life sciences [@lamy2019rainbio], health [@landolfi2022screening] and the retail industry [@adnan2018set], as well as in generic applications such as analysing structures of missing data [@ruddle2022using] and association rule mining [@wang2020visual]. SetVis is new matrix-based set visualization software, implemented as a Python package which is available from [PyPi](https://pypi.org/project/setvis). The documentation is available from [setvis.readthedocs.io](https://setvis.readthedocs.io) which contains various hands-on Jupyter notebooks guiding users to use the package. SetVis uses a memory-efficient design, operates with datasets held in RAM or a database such as PostgreSQL, and allows user interaction graphically or programmatically. A technical evaluation shows that SetVis uses orders of magnitude less memory than the UpSet [@lex2014upset] Python package [@nothman2022].
 
 # Statement of need
-Although a wide variety of set visualization software has been developed [@jia2021venn;@alsallakh2016powerset], most of such software generates Venn or Euler diagrams so is only suitable for data that contain fewer than 10 sets [@jia2021venn]. Other software visualizes 50+ sets but either has little support for set intersection tasks [@alper2011design;@dork2012pivotpaths;@kim2007visualizing;@freiler2008interactive] or only visualizes pairwise intersections [@molbiotools2022;@yalcin2015aggreset].
+Although a wide variety of set visualization software has been developed [@jia2021venn;@alsallakh2016powerset], most of such software generates Venn or Euler diagrams so is only suitable for data that contain fewer than ten sets [@jia2021venn]. Other software visualizes 50+ sets but either has little support for set intersection tasks [@alper2011design;@dork2012pivotpaths;@kim2007visualizing;@freiler2008interactive] or only visualizes pairwise intersections [@molbiotools2022;@yalcin2015aggreset].
 
 The best-known software for analysing rich patterns in set data are the R and Python UpSet plot packages [@conway2017upsetr;@nothman2022], but the memory requirement of both packages increases linearly with the number of cells (i.e., rows $\times$ columns) in a dataset, which makes the packages unusable with big data. The ACE software [@ruddle2022ace] uses more memory-efficient data structures, but first requires the whole of a dataset to have been loaded into RAM (again, clearly an issue for big data), and is a stand-alone Java application that cannot be integrated with Jupyter Notebooks or similar workflows. The SetVis python package addresses the above collective weaknesses because it: (a) operates with datasets that may be either held in RAM or out of core (in a PostgreSQL database), (b) stores sets and intersections in memory-efficient data structures (like ACE), (c) can be used within Jupyter Notebooks (or similar) to aid the replicability of analysis workflows, and (d) allows users to interact graphically in a notebook as well as programmatically.
 
 # Design
[…]
