### 2024-10-29 #6925 @mstimberg
@layik I've made a few suggestions for changes to the paper in alan-turing-institute/setvis#142, please merge the PR if you agree with them.

> The title is different as the package is called "SetVis" but the paper is called "SetVis: Visualizing Large Numbers of Sets and Intersections" everything else is checked and done.

Could you please use the paper title for the Zenodo archive as well – this is the recommended JOSS policy. Actually, I've noticed that the archive does not contain the code and result files for the benchmarks. It would be important to archive these as well, since otherwise there is no way to reproduce the figure from the paper.

Regarding the benchmark data and your earlier comment (apologies for not replying about this earlier!):
> Please see the files added to the joss branch of the repository. I intend to remove this from the repository once the review is concluded if you do not object to it.

Is there a reason to delete them? It is ok if you never merge the JOSS branch, so it wouldn't "pollute" your main branch. But of course, we cannot (and do not want to) control what you do with your GitHub repository after the publication ;-) But this again makes it important that the relevant data (in addition to the actual paper which we'll store and archive separately) is archived in a save place like Zenodo.
