### 2024-10-29 #6925 @layik
Thank you @mstimberg. Please find the new version (v2) with both title matching the paper, and the tech-evaluation files used by the reviewers within the release.
https://zenodo.org/records/14009355.

As for the `joss` branch, I internally suggested (since the v2 Zenodo release) that we only keep the .md + bib file within the repo and everything else is left in the Zenodo release.

I will respond to the PR separately which I appreciate!
