### 2024-10-29 90cb5bf
paper: minor clarification as per #142
