### 2024-10-30 #6925 @mstimberg
> Thank you @mstimberg. Please find the new version (v2) with both title matching the paper, and the tech-evaluation files used by the reviewers within the release. https://zenodo.org/records/14009355.
>
> As for the `joss` branch, I internally suggested (since the v2 Zenodo release) that we only keep the .md + bib file within the repo and everything else is left in the Zenodo release.

That all works for me, thanks for the updated archive. Just a minor thing: would you mind setting the version of the archive to `v0.1rc5` instead of `v2` (in the metadata, i.e. no new release/version necessary)? I don't think it is strictly necessary, but it might help avoid some confusion.
