### 2024-10-30 #6925 @layik
> setting the version of the archive to v0.1rc5 instead of v2
Done, thank you I had no idea this was possible.
