### 2024-10-30 #6925 @mstimberg
Thanks. Sorry about not noticing this earlier, but could you please update the `paper.bib` files with the suggested DOIs:

> - 10.1016/j.jclinepi.2020.12.004 may be a valid DOI for title: Methods for depicting overlap in overviews of syst...
> - 10.1093/ije/dyx015 may be a valid DOI for title: Data resource profile: Hospital episode statistics...
> - 10.1109/tvcg.2018.2864812 may be a valid DOI for title: Rulematrix: Visualizing and understanding classifi...
> - 10.1542/peds.2020-021634 may be a valid DOI for title: Pediatric palliative care programs in US hospitals

(Just to be clear: no need for new release/archive/version)
