### 2024-10-30 3fe0315
paper.bib: add doi to some refs
