### 2024-11-06 alan-turing-institute/setvis#135 @layik
Recommended to accept, can close this and leave #138 to finalise once paper is published.
